BUG COMMIT error building. seting NDK

+++ app/src/main/java/com/example/vlad/navigation/lenghtStep/alphaBettaAlgorithm/filters/Filter.java
@@ -10,5 +10,5 @@ package com.example.vlad.navigation.lenghtStep.alphaBettaAlgorithm.filters;
  * @author Roman
  */
 public interface Filter {
-    double[] filtering(double[] ar);
+    float[] filtering(float[] ar);
 }
